pyproject.toml
[tool.poetry]
name = "apis-instance-sicprod-dev"
version = "0.1.0"
description = "APIS Instance SiCProD"
authors = ["Birger Schacht <birger.schacht@oeaw.ac.at>"]
license = "MIT"
packages = [{include = "apis_ontology"}]


[tool.poetry.dependencies]
python = "^3.11"
apis-core = { git = "https://github.com/acdh-oeaw/apis-core-rdf.git", tag = "v0.7.0" }
apis-bibsonomy = { git = "https://github.com/acdh-oeaw/apis-bibsonomy.git", tag = "v0.8.0" }
apis-acdhch-default-settings = { git = "https://github.com/acdh-oeaw/apis-acdhch-default-settings.git", tag = "v0.1.10" }
psycopg2 = "^2.9.6"
django-matomo = "^0.1.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
